build(mypy): clear 4 physics validators (float64 dtype mismatch)
Sixth increment of the SYS-AUDIT-02-MYPY1 burn-down. Resolve every mypy
error in four physics-validation modules and remove them from the targeted
ignore list (43 modules remain, 85 now strict-checked).

The common cause is a dtype mismatch: np.linspace / profile arithmetic type
as floating[Any], while the src functions they feed declare
NDArray[np.float64]. np.linspace already returns float64 at runtime, so
`.astype(np.float64)` is a no-op that only reconciles the static type:
- validate_mhd_stability: shear/rho grids feeding _controlled_profile and
  QProfile (3 sites; np.full_like inherits the float64 dtype).
- validate_marfe_onset: the three MARFEFrontModel.T profile assignments.
- benchmark_hybrid_accuracy: the rho grid feeding GKScheduler.step /
  GKCorrector.update; run_hybrid_accuracy_benchmark -> dict[str, Any].
- validate_sawtooth_kadomtsev: the post-crash t_new/n_new profiles passed to
  the local _integral_rel_error(FloatArray) helper.

No behaviour change (astype float64 on already-float64 arrays). Pinned
pre-commit mypy passes; 66 validator tests (mhd, sawtooth, marfe) pass; no
optional third-party imports in these modules (CI env parity confirmed).

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)
Seat: c8f7

+++ validation/validate_marfe_onset.py
@@ -331,11 +331,11 @@ def scan_boundary_classification(config: MARFEConfig) -> ScanBoundaryClassificat
 def front_detection_thresholds() -> FrontDetectionThresholds:
     """Confirm the detached-front detector's two strict temperature thresholds."""
     below = MARFEFrontModel(L_par=100.0, kappa_par=20.0, q_perp=10.0, impurity="W", f_imp=1e-2)
-    below.T = np.linspace(19.0, 60.0, below.n_s)
+    below.T = np.linspace(19.0, 60.0, below.n_s).astype(np.float64)
     at_threshold = MARFEFrontModel(L_par=100.0, kappa_par=20.0, q_perp=10.0, impurity="W", f_imp=1e-2)
-    at_threshold.T = np.linspace(20.0, 60.0, at_threshold.n_s)
+    at_threshold.T = np.linspace(20.0, 60.0, at_threshold.n_s).astype(np.float64)
     hot = MARFEFrontModel(L_par=100.0, kappa_par=20.0, q_perp=10.0, impurity="W", f_imp=1e-2)
-    hot.T = np.linspace(30.0, 80.0, hot.n_s)
+    hot.T = np.linspace(30.0, 80.0, hot.n_s).astype(np.float64)
     return FrontDetectionThresholds(
         below_threshold_is_marfe=below.is_marfe(),
         at_temperature_threshold_is_marfe=at_threshold.is_marfe(),
